fix streamkey tests
